feat: add tf submodule hook
